requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
pydeck>=0.8.0
snowflake-connector-python>=3.7.0
requests>=2.31.0